requirements.txt
sentence-transformers
faiss-cpu
rank-bm25
transformers
wikipedia-api
beautifulsoup4
nltk
rouge-score
bert-score
scikit-learn
streamlit
torch
numpy
pandas
tqdm
sentencepiece
